pyproject.toml
[tool.poetry]
name = "email-api"
version = "0.1.0"
description = "A FastAPI backend for sending emails"
authors = ["You <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.110"
uvicorn = "^0.29"
aiosmtplib = "^2.0"
python-dotenv = "^1.0"
email-validator = "^2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
